Move 'flavorful' tests to their own test
