+++ crates/test-rust-wasm/src/bin/flavorful.rs
@@ -1,0 +1,3 @@
+include!("../../../../tests/runtime/flavorful/wasm.rs");
+
+fn main() {}
