+++ tests/runtime/flavorful/exports.witx
@@ -1,6 +1,4 @@
-// ===========================================
-// More flavorful types
-// ===========================================
+test_imports: function()
 
 record list_in_record1 { a: string }
 record list_in_record2 { a: string }
