+++ tests/runtime/flavorful/host.rs
@@ -1,19 +1,13 @@
-witx_bindgen_wasmtime::import!("tests/host.witx");
+use anyhow::Result;
 
-use host::*;
-pub(crate) use host::{add_host_to_linker, HostTables};
-use std::cell::RefCell;
-use witx_bindgen_wasmtime::{
-    imports::{PullBuffer, PushBuffer},
-    Le,
-};
+witx_bindgen_wasmtime::import!("./tests/runtime/flavorful/imports.witx");
+
+use imports::*;
 
 #[derive(Default)]
-pub struct MyHost {
-    scalar: u32,
-}
+pub struct MyImports;
 
-impl Host for MyHost {
+impl Imports for MyImports {
     fn list_in_record1(&mut self, ty: ListInRecord1<'_>) {
         assert_eq!(ty.a, "list_in_record1");
     }
@@ -95,4 +89,72 @@ impl Host for MyHost {
             vec![MyErrno::A, MyErrno::B],
         )
     }
+}
+
+witx_bindgen_wasmtime::export!("./tests/runtime/flavorful/exports.witx");
+
+fn run(wasm: &str) -> Result<()> {
+    use exports::*;
+
+    let (exports, mut store) = crate::instantiate(
+        wasm,
+        |linker| imports::add_imports_to_linker(linker, |cx| -> &mut MyImports { &mut cx.imports }),
+        |store, module, linker| Exports::instantiate(store, module, linker, |cx| &mut cx.exports),
+    )?;
+
+    exports.test_imports(&mut store)?;
+
+    exports.list_in_record1(
+        &mut store,
+        ListInRecord1 {
+            a: "list_in_record1",
+        },
+    )?;
+    assert_eq!(exports.list_in_record2(&mut store)?.a, "list_in_record2");
+
+    assert_eq!(
+        exports
+            .list_in_record3(
+                &mut store,
+                ListInRecord3Param {
+                    a: "list_in_record3 input"
+                }
+            )?
+            .a,
+        "list_in_record3 output"
+    );
+
+    assert_eq!(
+        exports
+            .list_in_record4(&mut store, ListInAliasParam { a: "input4" })?
+            .a,
+        "result4"
+    );
+
+    exports.list_in_variant1(
+        &mut store,
+        Some("foo"),
+        Err("bar"),
+        ListInVariant13::V0("baz"),
+    )?;
+    assert_eq!(
+        exports.list_in_variant2(&mut store)?,
+        Some("list_in_variant2".to_string())
+    );
+    assert_eq!(
+        exports.list_in_variant3(&mut store, Some("input3"))?,
+        Some("output3".to_string())
+    );
+
+    assert!(exports.errno_result(&mut store)?.is_err());
+    MyErrno::A.to_string();
+    format!("{:?}", MyErrno::A);
+    fn assert_error<T: std::error::Error>() {}
+    assert_error::<MyErrno>();
+
+    let (a, b) = exports.list_typedefs(&mut store, "typedef1", &["typedef2"])?;
+    assert_eq!(a, b"typedef3");
+    assert_eq!(b.len(), 1);
+    assert_eq!(b[0], "typedef4");
+    Ok(())
 }
