--- tests/runtime/rust-wasm/allocator.rs
@@ -1,1 +1,0 @@
-
