--- tests/runtime/rust-wasm/lib.rs
@@ -1,5 +1,0 @@
-#![cfg(target_arch = "wasm32")]
-
-mod allocator;
-mod exports;
-mod imports;
